requirements.txt
mysql-connector==2.2.9
openpyxl==3.1.4
pandas==2.2.2
